first steps ate database integration.

+++ roadmap/4-web-frameworks/1-basic/student/src/main/java/com/example/student/student/api/studentController.java
@@ -33,7 +33,7 @@ public class studentController {
     }
 
     @GetMapping("/getStudent/{id}")
-    public student getStudent(@PathVariable("id") int id) {           
+    public student getStudent(@PathVariable("id") Long id) {           
         return _studentService.getStudent(id);
     }
 }
